testing camera upload promises
