+++ src/pages/edit-profile/edit-profile.html
@@ -11,4 +11,12 @@
 
     <button ion-button color="primary" (click)="takeProfilePhoto()">Edit Photo</button>
 
+    Test:
+
+    {{ testvalue }}
+
+    ...
+
+    {{ testvalue | json }}
+
 </ion-content>
