+++ src/pages/edit-profile/edit-profile.ts
@@ -4,6 +4,7 @@ import { Camera, CameraOptions } from '@ionic-native/camera';
 import firebase from 'firebase';
 import { UserService } from '../../services/user.service';
 import { ErrorService } from '../../services/error.service';
+import { ImageService } from '../../services/image.service';
 
 @IonicPage()
 @Component({
@@ -16,7 +17,9 @@ export class EditProfilePage {
 
     private userImageRef: any;
 
-    constructor(private camera: Camera, private userService: UserService, private errorService: ErrorService) { 
+    testvalue:any = '';
+
+    constructor(private camera: Camera, private userService: UserService, private imageService: ImageService) { 
         this.userImageRef = firebase.storage().ref().child(`userprofile/${this.userService.activeUser.uid}.jpg`);
     }
 
@@ -33,42 +36,11 @@ export class EditProfilePage {
         });
     }
 
-    /*============================
-        Camera Photo and Upload
-    =============================*/
     takeProfilePhoto() {
-        const options: CameraOptions = {
-            quality: 90,
-            destinationType: this.camera.DestinationType.DATA_URL,
-            encodingType: this.camera.EncodingType.JPEG,
-            mediaType: this.camera.MediaType.PICTURE,
-            targetWidth: 576,
-            correctOrientation: true
-        }
-
-        this.camera.getPicture(options).then((imageData) => {
-            const imageBase64 = 'data:image/jpeg;base64,' + imageData;
-            this.upload(imageBase64);
-        })
-        .catch(err => {
-            if (!err.error.message) {
-                err.error = { message: 'There was a problem uploading. Please try again.' };
-            }
-
-            this.errorService.handleError(err);
+        this.imageService.takeProfilePhoto().then(res => {
+            this.testvalue = res;
+            this.profileImageUrl = res;
         })
     }
-
-    upload(imageBase64) {
-        this.userImageRef.putString(imageBase64, 'data_url').then(snapshot => {
-            this.profileImageUrl = snapshot.downloadURL;
-        })
-        .catch(err => {
-            if (!err.error.message) {
-                err.error = { message: 'There was a problem uploading. Please try again.' };
-            }
-
-            this.errorService.handleError(err);
-        });
-    }
+    
 }
